Add UI components and reorganize/shared types
Adds reusable UI controls for error display and mode toggling to promote consistent interfaces and reduce duplication across pages. Centralizes and splits domain types into focused modules with a single re-export surface to improve type discoverability and maintenance. Simplifies the XML upload hook by removing legacy per-file status tracking in favor of bin-based state and renames the clear function accordingly, reducing complexity in upload flow. Also introduces a typed API error shape to standardize error handling.

These changes make UI patterns reusable, streamline upload state logic, and improve overall type organization for easier future development and refactoring.
