+++ ui/src/components/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as ErrorBanner } from './ErrorBanner';
+export { default as ModeToggle } from './ModeToggle';
