+++ ui/src/hooks/useXmlUpload.ts
@@ -1,43 +1,11 @@
 import { useState } from 'react';
 import { catalogApi } from '../services/catalogApi';
 import { parseXmlToObservations, extractMetadataFromXml } from '../utils/xmlParser';
-import type { UploadStatus, MetadataExtractionRule, UploadBin } from '../types';
+import type { MetadataExtractionRule, UploadBin } from '../types';
 
 export const useXmlUpload = () => {
-  const [statuses, setStatuses] = useState<UploadStatus[]>([]);
   const [bins, setBins] = useState<UploadBin[]>([]);
   const [isScanning, setIsScanning] = useState(false);
-
-  // Legacy simple upload (kept for compatibility)
-  const uploadFiles = async (files: File[], contextId: string, metadata: Record<string, string>) => {
-    const initialStatuses: UploadStatus[] = files.map(file => ({
-      fileName: file.name,
-      status: 'pending'
-    }));
-    setStatuses(initialStatuses);
-
-    for (const file of files) {
-      updateStatus(file.name, { status: 'parsing' });
-
-      try {
-        const content = await file.text();
-        const observations = parseXmlToObservations(content, metadata);
-
-        updateStatus(file.name, { status: 'submitting', observationCount: observations.length });
-
-        await catalogApi.submitObservations(contextId, observations);
-
-        updateStatus(file.name, { status: 'complete' });
-      } catch (error) {
-        console.error(`Failed to upload ${file.name}`, error);
-        updateStatus(file.name, { status: 'error', error: (error as Error).message });
-      }
-    }
-  };
-
-  const updateStatus = (fileName: string, update: Partial<UploadStatus>) => {
-    setStatuses(prev => prev.map(s => s.fileName === fileName ? { ...s, ...update } : s));
-  };
 
   const scanFiles = async (files: File[], rules: Record<string, MetadataExtractionRule>) => {
     setIsScanning(true);
@@ -105,20 +73,17 @@ export const useXmlUpload = () => {
     }
   };
 
-  const clearStatuses = () => {
-    setStatuses([]);
+  const clearBins = () => {
     setBins([]);
   };
 
   return {
-    uploadFiles,
-    statuses,
     bins,
     scanFiles,
     updateBinMetadata,
     submitBin,
-    clearStatuses,
-    isUploading: statuses.some(s => s.status === 'parsing' || s.status === 'submitting') || bins.some(b => b.status === 'submitting'),
+    clearBins,
+    isUploading: bins.some(b => b.status === 'submitting'),
     isScanning
   };
 };
