+++ ui/src/types/api.types.ts
@@ -1,0 +1,7 @@
+export interface ErrorResponse {
+  message: string;
+  status: number;
+  timestamp: string;
+  error: string;
+  errors?: string[];
+}
