+++ ui/src/types/index.ts
@@ -1,110 +1,28 @@
-export interface MetadataExtractionRule {
-  xpaths: string[];
-  validationRegex?: string;
-}
+// Context types
+export type {
+  MetadataExtractionRule,
+  Context,
+  ContextWithCount,
+} from './context.types';
 
-export interface Context {
-  contextId: string;
-  displayName: string;
-  description: string | null;
-  requiredMetadata: string[];
-  optionalMetadata: string[];
-  metadataRules: Record<string, MetadataExtractionRule>;
-  active: boolean;
-  createdAt: string;
-  updatedAt: string | null;
-}
+// Catalog types
+export type {
+  CatalogSearchRequest,
+  CatalogEntry,
+  CatalogObservation,
+  PagedResponse,
+} from './catalog.types';
 
-export interface ContextWithCount extends Context {
-  fieldCount: number;
-}
+// Facet types
+export type {
+  FacetValue,
+  FacetState,
+  FacetIndex,
+  UseFacetsReturn,
+} from './facet.types';
 
-export interface CatalogSearchRequest {
-  q?: string;
-  contextId?: string;
-  fieldPathContains?: string;
-  metadata?: Record<string, string>;
-  page?: number;
-  size?: number;
-  sort?: string;
-  useRegex?: boolean;
-}
+// Upload types
+export type { UploadBin } from './upload.types';
 
-export interface CatalogEntry {
-  id: string;
-  contextId: string;
-  metadata: Record<string, string>;
-  fieldPath: string;
-  maxOccurs: number;
-  minOccurs: number;
-  allowsNull: boolean;
-  allowsEmpty: boolean;
-  firstObservedAt: string;
-  lastObservedAt: string;
-}
-
-export interface CatalogObservation {
-  metadata: Record<string, string>;
-  fieldPath: string;
-  count: number;
-  hasNull: boolean;
-  hasEmpty: boolean;
-}
-
-export interface PagedResponse<T> {
-  content: T[];
-  totalElements: number;
-  totalPages: number;
-  size: number;
-  number: number;
-  first: boolean;
-  last: boolean;
-}
-
-export interface UploadStatus {
-  fileName: string;
-  status: 'pending' | 'parsing' | 'submitting' | 'complete' | 'error';
-  observationCount?: number;
-  error?: string;
-}
-
-export interface ErrorResponse {
-  message: string;
-  status: number;
-  timestamp: string;
-  error: string;
-  errors?: string[];
-}
-
-export interface FacetValue {
-  value: string;
-  count: number;
-}
-
-export interface FacetState {
-  values: FacetValue[];
-  mode: 'any' | 'one';
-  selected: Set<string>;
-}
-
-export interface FacetIndex {
-  [key: string]: FacetState;
-}
-
-export interface UploadBin {
-  id: string;
-  files: File[];
-  metadata: Record<string, string>;
-  status: 'pending' | 'submitting' | 'complete' | 'error';
-  error?: string;
-  progress: number;
-}
-
-export interface useFacetsReturn {
-  facets: FacetIndex;
-  filteredResults: CatalogEntry[];
-  setFacetMode: (key: string, mode: 'any' | 'one') => void;
-  toggleFacetValue: (key: string, value: string) => void;
-  clearFacet: (key: string) => void;
-  clearAllFacets: () => void;
-}
+// API types
+export type { ErrorResponse } from './api.types';
